Updated functions and uses
Still messing around with the code here to see what works.  As of right
now, this works with two different json files, one being local and the
other being remote.  When we write data, it goes to local, when we read
data, it combines the data from both local and remote and then returns
that data.