Made some fundamental changes to code
Created a custom version of the edit-json-file library to be able to
sync with a file after it has been downloaded by gdrive.  For what ever
reason, the default library would return blank after a download.

Revised the Model functions to be more generic so we can have a bit more
flexibility inside the index.js Client file. We want these all to be as
generic as possible.  The only node that should be doing heavy lifting
is the front-end nodes. This allows for broader flexibility and less
updating of back-end nodes.